site: Add a primary color change button
